Matrix - Start
With HPC at the University of Delaware, we went though weeks of research to ultimately understand how to parallelize code on a GPU. To test my knowledge, I used matrix multiplication for practice. This is the initial code I used with no parallelization involved. It is simple and uses a few for loops to accomplish the task.